refactor(hypav3): extract ModalFooter
