+++ src/lib/Others/HypaV3Modal/types.ts
@@ -36,3 +36,11 @@ export interface ChatMemoSearchResult {
   summaryIndex: number;
   memoIndex: number;
 }
+
+export const DISPLAY_MODE = {
+  All: "All",
+  Range: "Range",
+  Recent: "Recent",
+} as const;
+
+export type DisplayMode = (typeof DISPLAY_MODE)[keyof typeof DISPLAY_MODE];
